Resolve conflicts and fix tests
