+++ setup.py
@@ -9,7 +9,7 @@ setup(
     classifiers=["Programming Language :: Python :: 3 :: Only"],
     py_modules=["tap_appsflyer"],
     install_requires=[
-        "attrs==25.1.0",
+        "attrs==26.1.0",
         "singer-python==6.8.0",
         "requests==2.34.2",
         "backoff==2.2.1",
